% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.r
\name{compile.sscr}
\alias{compile.sscr}
\title{Compiling TMB C++ templates}
\usage{
compile.sscr()
}
\description{
Compiles the \code{sscr} TMB templates into a shared object
file. This must be done a single time following installation or
updating of the package.
}

