% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-sscp.r
\name{fit.sscr}
\alias{fit.sscr}
\title{Fitting an SCR model with second-order spatial dependence}
\usage{
fit.sscr(capt, traps, mask, resp = "binom", re.structure = "none")
}
\arguments{
\item{capt}{Capture history object.}

\item{traps}{Traps object.}

\item{mask}{Mask object.}

\item{resp}{Response distribution for capture history elements.}

\item{re.structure}{Structure of}
}
\description{
Fits an SSCR model. Estimation is by maximum likelihood. The
second-order spatial dependence is modelled via trap-level random
effects for each detected individual. The likelihood function is
calculated by integrating over these random effects using the
Laplace approximation.
}

