% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\name{sim.sscr}
\alias{sim.sscr}
\title{Simulating SCR data with second-order spatial dependence}
\usage{
sim.sscr(traps, mask, D, resp, resp.pars, detfn = "hn", detfn.scale = "er",
  cov.structure, re.scale = "er", det.pars = NULL, cov.pars = NULL)
}
\arguments{
\item{traps}{A matrix with two columns, providing the Cartesian
coordinates of the detector locations.}

\item{mask}{A mask object for integration over the survey area.}

\item{D}{Animal density (individuals per hectare).}

\item{resp}{Response distribution for capture history
elements. Either \code{"binom"} for a binomial distribution, or
\code{"pois"} for a Poisson distribution.}

\item{resp.pars}{A named vector of known, fixed parameters for the
response distribution. If \code{resp} is \code{"binom"}, then
this must have a single element named \code{"size"} giving the
fixed number of trials; if this argument is not provided, then
the default is 1.}

\item{detfn}{Detection function, given by a character string. Use
\code{"hn"} for halfnormal and \code{"hr"} for hazard rate.}

\item{detfn.scale}{A character string, either \code{"er"} or
\code{"prob"}. This indicates whether the detection function
should provide the encounter rate (expected number of
detections) or the probability of detection.}

\item{cov.structure}{Covariance structure of the random
effects. The current options are (1) \code{"none"} for no
random effects (regular SCR), (2) \code{"independent"}, for
independent random effects (equivalent to counts of detections
being overdispersed), (3) \code{"exponential"}, for random
effects with an exponential covariance structure, (4)
"sq_exponential" for random effects with a squared exponential
covariance structure, (5) \code{"matern"}, for random effects
with a Matern covariance structure, (5) \code{"individual"},
for random effects that are restricted to being the same at all
traps (equivalent to having an independent random effect on
\code{lambda0} for each individual), or (6)
\code{"lc_exponential"} for a linear combination of exponential
covariance functions.}

\item{re.scale}{A character string, either \code{"er"} or
\code{"prob"}. This indicates whether the Gaussian random
effects effect the encounter rate (expected number of
detections) or the probability of detection.}

\item{det.pars}{List of detection function parameters.}

\item{cov.pars}{List of covariance parameters.}
}
\description{
Simulates SSCR data.
}
