% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-sscr.r
\name{fit.sscr}
\alias{fit.sscr}
\title{Fitting an SCR model with second-order spatial dependence}
\usage{
fit.sscr(
  capt,
  traps,
  mask,
  resp = "binom",
  resp.pars = NULL,
  detfn = "hhn",
  cov.structure = "none",
  re.multiplier = "er",
  start = NULL,
  fix = NULL,
  toa = NULL,
  trace = FALSE,
  test = FALSE,
  test.conditional.n = TRUE,
  hess = FALSE,
  exact.gr = TRUE,
  optim.fun = "nlminb"
)
}
\arguments{
\item{capt}{A capture history object. It should be a matrix, where
the jth element of the ith row should provide a detection
record of the ith individual at the jth detector.}

\item{traps}{A matrix with two columns, providing the Cartesian
coordinates of the detector locations.}

\item{mask}{A mask object for integration over the survey area.}

\item{resp}{Response distribution for capture history
elements. Either \code{"binom"} for a binomial distribution,
\code{"pois"} for a Poisson distribution, \code{"nb"} for a
negative binomial distribution, \code{"nba"} for a negative
binomial--alpha distribution, or \code{"cmp"} for a
Conway-Maxwell-Poisson distribution.}

\item{resp.pars}{A named vector of additional parameters for the
response distribution. If \code{resp} is \code{"binom"}, then
this should have a single element named giving the fixed number
of trials. If \code{resp} is \code{"cmp"}, then this should
have a single element for the start value of the estimated
parameter \code{nu} (see \link{rcmp}). If \code{resp} is
\code{"nb"}, then this should have a single element for the
start value of the estimated parameter \code{size} (see
\link{rnbinom}).}

\item{detfn}{Detection function, given by a character string. Use
\code{"hn"}, for halfnormal, \code{"hhn"} for hazard
halfnormal, and \code{"hr"} for hazard rate.}

\item{cov.structure}{Covariance structure of the random
effects. The current options are (1) \code{"none"} for no
random effects (regular SCR), (2) \code{"independent"}, for
independent random effects (equivalent to counts of detections
being overdispersed), (3) \code{"exponential"}, for random
effects with an exponential covariance structure, (4)
\code{"sq_exponential"} for random effects with a squared
exponential covariance structure, (5) \code{"matern"}, for
random effects with a Matern covariance structure, (5)
\code{"individual"}, for random effects that are restricted to
being the same at all traps (equivalent to having an
independent random effect on \code{lambda0} for each
individual), or (6) \code{"lc_exponential"} for a linear
combination of exponential covariance functions.}

\item{re.multiplier}{A character string specifying how the expected
encounter rate is calculated from the random effect. This is
either (1) \code{"er"} for multiplication by the baseline
encounter rate, or (2) \code{"prob"} for multiplication by the
baseline encounter probability.}

\item{start}{A named vector of parameter start values.}

\item{fix}{A named vector of parameter values that are fixed in the
model. Supercedes values provided in \code{start}.}

\item{toa}{A matrix with the same dimensions as \code{capt} that
provides time-of-arrival information for acoustic detections.}

\item{trace}{Logical. If \code{TRUE}, parameter values for each
step of the optimisation algorithm are printed.}

\item{test}{Logical. If \code{TRUE}, the negative log-likelihood is
calculated at parameter start values. If \code{FALSE}, a model
is fitted. Alternatively, a character string. If \code{"nll"},
then the negative log-likelihood is calculated. If \code{"gr"},
then the partial derivatives of the negative log-likelihood
function with respect to the parameters is also calculated. If
\code{"hess"} then the Hessian if also calculated.}

\item{test.conditional.n}{Logical. If \code{TRUE}, tests are
carried out conditioning in the number of detections.}

\item{hess}{Logical. If \code{TRUE}, a Hessian is computed and a
variance-covariance matrix is returned.}

\item{exact.gr}{Logical. If \code{TRUE}, partial derivatives of the
likelihood with respect to parameters are calculated
analytically via automatic differentiation. If \code{FALSE}
they are approximated numerically.}

\item{optim.fun}{A character string representing the R function to
maximise the likelihood. This can be \code{"bobyqa"} (from
package \code{minqa}), \code{"nlminb"}, or \code{"nlm"}.}
}
\description{
Fits an SSCR model. Estimation is by maximum likelihood. The
second-order spatial dependence is modelled via trap-level random
effects for each detected individual. The likelihood function is
calculated by integrating over these random effects using the
Laplace approximation.
}
