% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-sscp.r
\name{fit.sscr}
\alias{fit.sscr}
\title{Fitting an SCR model with second-order spatial dependence}
\usage{
fit.sscr(capt, traps, mask, resp = "binom", re.structure = "none",
  trace = FALSE, test = FALSE)
}
\arguments{
\item{capt}{Capture history object.}

\item{traps}{Traps object.}

\item{mask}{Mask object.}

\item{resp}{Response distribution for capture history elements.}

\item{re.structure}{Covariance structure of the random effects. The
current options are (1) \code{"none"} for no random effects
(regular SCR), (2) \code{"independent"}, for independent random
effects (equivalent to counts of detections being
overdispersed), (3) \code{"constant"}, for independent random
effects that are restricted to being the same at all traps
(equivalent to having a random effect on \code{lambda0} for
each individual), and (4) \code{"exponential"}, for random
effects with an exponential covariance structure (spatially
structured random effects).}

\item{trace}{Logical. If \code{TRUE}, parameter values for each
step of the optimisation algorithm are printed.}

\item{test}{Logical. If \code{TRUE}, the likelihood is calculated
at parameter start values. If \code{FALSE}, a model is fitted.}
}
\description{
Fits an SSCR model. Estimation is by maximum likelihood. The
second-order spatial dependence is modelled via trap-level random
effects for each detected individual. The likelihood function is
calculated by integrating over these random effects using the
Laplace approximation.
}

