% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-sscr.r
\name{fit.sscr}
\alias{fit.sscr}
\title{Fitting an SCR model with second-order spatial dependence}
\usage{
fit.sscr(capt, traps, mask, resp = "binom", resp.pars = NULL,
  detfn = "hn", cov.structure = "none", trace = FALSE, test = FALSE)
}
\arguments{
\item{capt}{Capture history object.}

\item{traps}{Traps object.}

\item{mask}{Mask object.}

\item{resp}{Response distribution for capture history
elements. Either \code{"binom"} for a binomial distribution, or
\code{"pois"} for a Poisson distribution.}

\item{resp.pars}{A named vector of known, fixed parameters for the
response distribution. If \code{resp} is \code{"binom"}, then
this must have a single element named \code{"size"} giving the
fixed number of trials; if this argument is not provided, then
the default is 1.}

\item{detfn}{Detection function, given by a character string. Use
\code{"hn"} for halfnormal and \code{"hr"} for hazard rate.}

\item{cov.structure}{Covariance structure of the random
effects. The current options are (1) \code{"none"} for no
random effects (regular SCR), (2) \code{"independent"}, for
independent random effects (equivalent to counts of detections
being overdispersed), (3) \code{"exponential"}, for random
effects with an exponential covariance structure, (4)
"sq_exponential" ofr random effects with a squared exponential
covariance structure, (5) \code{"matern"}, for random effects
with a Matern covariance structure, (5) \code{"individual"},
for random effects that are restricted to being the same at all
traps (equivalent to having an independent random effect on
\code{lambda0} for each individual), or (6)
\code{"lc_exponential"} for a linear combination of exponential
covariance functions.}

\item{trace}{Logical. If \code{TRUE}, parameter values for each
step of the optimisation algorithm are printed.}

\item{test}{Logical. If \code{TRUE}, the likelihood is calculated
at parameter start values. If \code{FALSE}, a model is fitted.}
}
\description{
Fits an SSCR model. Estimation is by maximum likelihood. The
second-order spatial dependence is modelled via trap-level random
effects for each detected individual. The likelihood function is
calculated by integrating over these random effects using the
Laplace approximation.
}

