% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-sscr.r
\name{fit.sscr}
\alias{fit.sscr}
\title{Fitting an SCR model with second-order spatial dependence}
\usage{
fit.sscr(capt, traps, mask, resp = "binom", resp.pars = NULL,
  detfn = "hn", detfn.scale = "er", cov.structure = "none",
  re.scale = "er", start = NULL, toa = NULL, trace = FALSE,
  test = FALSE, hess = FALSE)
}
\arguments{
\item{capt}{A capture history object. It should be a matrix, where
the jth element of the ith row should provide a detection
record of the ith individual at the jth detector.}

\item{traps}{A matrix with two columns, providing the Cartesian
coordinates of the detector locations.}

\item{mask}{A mask object for integration over the survey area.}

\item{resp}{Response distribution for capture history
elements. Either \code{"binom"} for a binomial distribution, or
\code{"pois"} for a Poisson distribution.}

\item{resp.pars}{A named vector of known, fixed parameters for the
response distribution. If \code{resp} is \code{"binom"}, then
this must have a single element named \code{"size"} giving the
fixed number of trials; if this argument is not provided, then
the default is 1.}

\item{detfn}{Detection function, given by a character string. Use
\code{"hn"} for halfnormal and \code{"hr"} for hazard rate.}

\item{detfn.scale}{A character string, either \code{"er"} or
\code{"prob"}. This indicates whether the detection function
should provide the encounter rate (expected number of
detections) or the probability of detection.}

\item{cov.structure}{Covariance structure of the random
effects. The current options are (1) \code{"none"} for no
random effects (regular SCR), (2) \code{"independent"}, for
independent random effects (equivalent to counts of detections
being overdispersed), (3) \code{"exponential"}, for random
effects with an exponential covariance structure, (4)
"sq_exponential" for random effects with a squared exponential
covariance structure, (5) \code{"matern"}, for random effects
with a Matern covariance structure, (5) \code{"individual"},
for random effects that are restricted to being the same at all
traps (equivalent to having an independent random effect on
\code{lambda0} for each individual), or (6)
\code{"lc_exponential"} for a linear combination of exponential
covariance functions.}

\item{re.scale}{A character string, either \code{"er"} or
\code{"prob"}. This indicates whether the Gaussian random
effects effect the encounter rate (expected number of
detections) or the probability of detection.}

\item{start}{A named list of parameter start values.}

\item{toa}{A matrix with the same dimensions as \code{capt} that
provides time-of-arrival information for acoustic detections.}

\item{trace}{Logical. If \code{TRUE}, parameter values for each
step of the optimisation algorithm are printed.}

\item{test}{Logical. If \code{TRUE}, the likelihood is calculated
at parameter start values. If \code{FALSE}, a model is fitted.}
}
\description{
Fits an SSCR model. Estimation is by maximum likelihood. The
second-order spatial dependence is modelled via trap-level random
effects for each detected individual. The likelihood function is
calculated by integrating over these random effects using the
Laplace approximation.
}
