% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\name{sim.sscr}
\alias{sim.sscr}
\title{Simulating SCR data with second-order spatial dependence}
\usage{
sim.sscr(fit = NULL, traps, mask, D, resp = NULL, resp.pars = NULL,
  detfn = "hhn", cov.structure = "none", re.multiplier = "er",
  det.pars = NULL, cov.pars = NULL, toa.pars = NULL,
  trunc.acs = TRUE, keep.acs = FALSE, mask.spacing = NULL)
}
\arguments{
\item{fit}{A fitted model object returned by \link{fit.sscr}(). If
provided, simulation is based on this fitted model, and all
other arguments are ignored.}

\item{traps}{A matrix with two columns, providing the Cartesian
coordinates of the detector locations.}

\item{mask}{A mask object for integration over the survey area.}

\item{D}{Animal density (individuals per hectare).}

\item{resp}{Response distribution for capture history
elements. Either \code{"binom"} for a binomial distribution,
\code{"pois"} for a Poisson distribution, or \code{"cmp"} for a
Conway-Maxwell-Poisson distribution.}

\item{resp.pars}{A named vector of known, fixed parameters for the
response distribution. If \code{resp} is \code{"binom"}, then
this must have a single element named \code{"size"} giving the
fixed number of trials; if this argument is not provided, then
the default is 1.}

\item{detfn}{Detection function, given by a character string. Use
\code{"hn"}, for halfnormal, \code{"hhn"} for hazard
halfnormal, and \code{"hr"} for hazard rate.}

\item{cov.structure}{Covariance structure of the random
effects. The current options are (1) \code{"none"} for no
random effects (regular SCR), (2) \code{"independent"}, for
independent random effects (equivalent to counts of detections
being overdispersed), (3) \code{"exponential"}, for random
effects with an exponential covariance structure, (4)
\code{"sq_exponential"} for random effects with a squared
exponential covariance structure, (5) \code{"matern"}, for
random effects with a Matern covariance structure, (5)
\code{"individual"}, for random effects that are restricted to
being the same at all traps (equivalent to having an
independent random effect on \code{lambda0} for each
individual), or (6) \code{"lc_exponential"} for a linear
combination of exponential covariance functions.}

\item{re.multiplier}{A character string specifying how the expected
encounter rate is calculated from the random effect. This is
either (1) \code{"er"} for multiplication by the baseline
encounter rate, or (2) \code{"prob"} for multiplication by the
baseline encounter probability.}

\item{det.pars}{List of detection function parameters.}

\item{cov.pars}{List of covariance parameters.}

\item{toa.pars}{List of time-of-arrival parameters.}

\item{trunc.acs}{Logical. If \code{TRUE}, activity centres are
truncated to the region covered by the mask.}

\item{keep.acs}{Logical. If \code{TRUE}, activity centre locations
are returned along with the captures histories.}

\item{mask.spacing}{Optional. The distance between adjacent mask
points. Providing a value will save processing time if activity
centres are being truncated to the mask region.}
}
\description{
Simulates SSCR data.
}
