% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\name{sim.sscr}
\alias{sim.sscr}
\title{Simulating SCR data with second-order spatial dependence}
\usage{
sim.sscr(traps, mask, resp, cov.structure, D, det.pars = NULL,
  cov.pars = NULL)
}
\arguments{
\item{traps}{Traps object.}

\item{mask}{Mask object.}

\item{resp}{Response distribution for capture history elements.}

\item{cov.structure}{Covariance structure of the random
effects. The current options are (1) \code{"none"} for no
random effects (regular SCR), (2) \code{"independent"}, for
independent random effects (equivalent to counts of detections
being overdispersed), (3) \code{"exponential"}, for random
effects with an exponential covariance structure, (4)
\code{"matern"}, for random effects with a Matern covariance
structure, (5) \code{"individual"}, for random effects that are
restricted to being the same at all traps (equivalent to having
an independent random effect on \code{lambda0} for each
individual), or (6) \code{"lc_exponential"} for a linear
combination of exponential covariance functions.}

\item{D}{Animal density (individuals per hectare).}

\item{det.pars}{List of detection function parameters.}

\item{cov.pars}{List of covariance parameters.}
}
\description{
Simulates SSCR data.
}

